+++ src/game/client/tf/player_vs_environment/c_tf_upgrades.h
@@ -82,7 +82,7 @@ public:
 
 	bool CheckInspectMode( void ) {
 		CSteamID steamID;
-		return m_bInspectMode && m_hPlayer->GetSteamID( &steamID ) != false;
+		return m_bInspectMode && m_hPlayer && m_hPlayer->GetSteamID( &steamID ) != false;
 	};
 
 public: